+++ i18n/ads-language-pack-pt-BR/CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to the "ads-language-pack-pt_BR" language pack will be documented in this file.
 
 ## [Released]
+* April 6, 2022 - Release for Azure Data Studio 1.36
 * February 23, 2022 - Release for Azure Data Studio 1.35
 * December 14, 2021 - Release for Azure Data Studio 1.34
 * May 21, 2021 - Release for Azure Data Studio 1.29